agent-classifier-chat-app/my-chat-app/requirements.txt
azure-ai-projects
azure-identity
semantic-kernel
python-dotenv
markdown
aiohttp
asyncio